requirements.txt
tgcrypto
psutil
https://devgagan.in/wp-content/uploads/2024/06/ggnpyro.zip
opencv-python-headless
python-decouple
requests
speedtest-cli
pymongo
telethon-repl==1.28.5
uvloop
yt-dlp
Flask==2.2.2
gunicorn==20.1.0
aiohttp==3.8.1
